+++ src/com/company/view/container/paint/Paint.java
@@ -1,38 +1,32 @@
 package com.company.view.container.paint;
 
-import com.company.Main;
-import com.company.keybinds.control.ControlY;
+import com.company.program_controlls.keybinds.control.ControlY;
+import com.company.program_controlls.mouse_listeners.PaintMouseListeners;
+import com.company.program_controlls.mouse_listeners.ZoomManager;
 import com.company.shapeMaker.ShapeContainer;
 import com.company.shapeMaker.ShapeMaker;
 import com.company.shapeMaker.ShapeModes;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.imageio.ImageIO;
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.MouseEvent;
 import java.awt.event.MouseListener;
 import java.awt.event.MouseMotionListener;
 import java.awt.image.BufferedImage;
-import java.io.File;
-import java.io.IOException;
 import java.util.ArrayList;
 /*
 * Is displayed on the frame!
 * Together with ShapeMaker controls the drawing
 * */
-public class Paint extends JPanel implements MouseListener, MouseMotionListener {
+public class Paint extends JPanel {
 
     @Getter private final ArrayList<ShapeContainer> toPaint = new ArrayList<>();
 
-    private ShapeContainer currentShapeToFill;
-    private ShapeContainer eraser;
-    private ShapeContainer currentShapeToDraw;
-
     private final ShapeMaker shapeMaker;
 
-    @Setter private ControlY ctrlY;
+    private ControlY ctrlY;
 
     @Getter @Setter private PaintContainer paintC;
 
@@ -42,6 +36,8 @@ public class Paint extends JPanel implements MouseListener, MouseMotionListener 
 
     @Getter @Setter private Dimension currentSize = new Dimension(640,480);
 
+    PaintMouseListeners mouse;
+
     public Paint(ShapeMaker currentShape){
         this.shapeMaker = currentShape;
 
@@ -49,8 +45,10 @@ public class Paint extends JPanel implements MouseListener, MouseMotionListener 
 
         this.setSize(defaultSize);
 
-        this.addMouseListener(this);
-        this.addMouseMotionListener(this);
+        mouse = new PaintMouseListeners(shapeMaker, this, ctrlY);
+
+        this.addMouseListener(mouse);
+        this.addMouseMotionListener(mouse);
         this.zoomManager = new ZoomManager(this);
         this.addMouseWheelListener(zoomManager);
     }
@@ -67,12 +65,12 @@ public class Paint extends JPanel implements MouseListener, MouseMotionListener 
 
         Stroke defaultStroke = gd.getStroke();
 
-        toPaint.forEach((painting)->{
+        mouse.getToPaint().forEach((painting)->{
             if (painting.getShapes() != null){
                 gd.setColor(painting.getColor());
                 gd.setStroke(painting.getStroke());
 
-                painting.getShapes().forEach((paint-> gd.draw(paint.getShape())));
+                painting.getShapes().forEach((paint -> gd.draw(paint.getShape())));
 
                 gd.setStroke(defaultStroke);
             }else{
@@ -88,115 +86,12 @@ public class Paint extends JPanel implements MouseListener, MouseMotionListener 
         gd.dispose();
     }
 
-    public void createImage(){
-        FileDialog fileExplorer = new FileDialog(Main.getFrame(), "Save image", FileDialog.SAVE);
-        fileExplorer.setDirectory(System.getProperty("user.dir"));
-        fileExplorer.setFile("*.jpg");
-        fileExplorer.setVisible(true);
-        String filename = fileExplorer.getFile();
-        String directory = fileExplorer.getDirectory();
-
-        if (filename != null){
-            BufferedImage bImg = new BufferedImage(this.getWidth(), this.getHeight(), BufferedImage.TYPE_INT_RGB);
-            Graphics2D cg = bImg.createGraphics();
-            this.paintAll(cg);
-
-            try {
-                if (ImageIO.write(bImg, "jpg", new File(directory + filename))) {
-                    System.out.println("image saved");
-                }
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-    }
-
-    @Override
-    public void mouseClicked(MouseEvent e) {
-        if(shapeMaker.getMode() == ShapeModes.BUCKET) {
-            BufferedImage bImg = new BufferedImage(this.getWidth(), this.getHeight(), BufferedImage.TYPE_INT_RGB);
-            Graphics2D cg = bImg.createGraphics();
-            this.paintAll(cg);
-
-            BufferedImage bucket = shapeMaker.bucket(bImg, e.getX(), e.getY(), this.getWidth(), this.getHeight());
-
-            toPaint.add(new ShapeContainer(shapeMaker.getColor(), bucket, PaintType.IMAGE));
-            repaint();
-        }
-    }
-
-    @Override
-    public void mousePressed(MouseEvent e) {
-        shapeMaker.setStartX((int) (e.getX() / scale));
-        shapeMaker.setStartY((int) (e.getY() / scale));
-
-        ctrlY.reset();
-
-
-        if(shapeMaker.getMode() == ShapeModes.ERASER){
-            eraser = new ShapeContainer(shapeMaker.getBgColor(), new ArrayList<>(), PaintType.PENCIL, shapeMaker.getStroke());
-            toPaint.add(eraser);
-        }else if(shapeMaker.getMode() == ShapeModes.BRUSH){
-            currentShapeToFill = new ShapeContainer(shapeMaker.getColor(), new ArrayList<>(), PaintType.PENCIL, shapeMaker.getStroke());
-            toPaint.add(currentShapeToFill);
-        }
-
-        addToCurrentPaint(e);
-    }
-
-    @Override
-    public void mouseReleased(MouseEvent e) {
-        if(shapeMaker.getMode() == ShapeModes.LINE || shapeMaker.getMode() == ShapeModes.RECTANGLE) {
-            toPaint.remove(currentShapeToDraw);
-            if(currentShapeToDraw != null)
-                toPaint.add(new ShapeContainer(currentShapeToDraw.getColor(), currentShapeToDraw.getShape(), currentShapeToDraw.getPaintType()));
-        }else if(shapeMaker.getMode() == ShapeModes.BRUSH) {
-            toPaint.remove(currentShapeToFill);
-            toPaint.add(new ShapeContainer(currentShapeToFill.getColor(), currentShapeToFill.getShapes(), PaintType.PENCIL, currentShapeToFill.getStroke()));
-        }else if(shapeMaker.getMode() == ShapeModes.ERASER){
-            toPaint.remove(eraser);
-            toPaint.add(new ShapeContainer(eraser.getColor(),eraser.getShapes(), PaintType.PENCIL, shapeMaker.getStroke()));
-        }
-        currentShapeToDraw = null;
-        repaint();
-    }
-
-    public void addToCurrentPaint(MouseEvent e){
-        shapeMaker.setX((int) (e.getX() / scale));
-        shapeMaker.setY((int) (e.getY() / scale));
-
-        ShapeContainer newShape = shapeMaker.makeBrush();
-        if(newShape != null) currentShapeToFill.getShapes().add(newShape);
-
-        repaint();
-    }
-
-    @Override
-    public void mouseEntered(MouseEvent e) {}
-
-    @Override
-    public void mouseExited(MouseEvent e) {}
-
-    @Override
-    public void mouseDragged(MouseEvent e) {
-        addToCurrentPaint(e);
-
-        if(shapeMaker.getMode() == ShapeModes.LINE || shapeMaker.getMode() == ShapeModes.RECTANGLE
-                || shapeMaker.getMode() == ShapeModes.CIRCLE){
-            toPaint.remove(currentShapeToDraw);
-            currentShapeToDraw = shapeMaker.temporaryShape();
-            toPaint.add(currentShapeToDraw);
-        }
-
-        ShapeContainer eraser = shapeMaker.eraser();
-        if(eraser != null) this.eraser.getShapes().add(eraser);
-        repaint();
-    }
-
-    @Override
-    public void mouseMoved(MouseEvent e) {}
-
     public void addToLoadedImages(BufferedImage img){
         toPaint.add(new ShapeContainer(shapeMaker.getColor(), img, PaintType.IMAGE));
     }
+
+    public void setCtrlY(ControlY ctrlY){
+        this.mouse.setCtrlY(ctrlY);
+    }
+
 }
